@hanhanW 2021-11-30
> The DeeplabV3 and MobileSSD regressions seem real?

I don't think so. It looks more like noisy on S20. The ARM configuration should be the same. I triggered the benchmark build to verify it. I'll double check with my local Pixel 4.

> L1 tile sizes maybe still too big? I would expect 2, 32, 32 to do well

I tried `2, 32, 32` and `4, 32, 32` but it's worse. I haven't looked into details yet because I think we can tune it later.